Primary Commit of Project Snowshoes
Welcome, friends, family, colleagues, dogs that run around the studio,
and random prowling bots! This is the first commit for Project
Snowshoes, a project that previously had its code privately held for a
year and a half. Now, anything that gets done can actually be seen here.

A little note: just because this is going live, the craziness might not
stop. If you look through the code, you might see that some of the
method names and whatnot are absolutely disgusting. I know...I look back
at them as well, and they are difficult to interpret. While I am
actually compiling a document to decode some of the more confusing names
and references used in method/instance names, etc., things will get
better now that this moved on from something I do at three in the
morning to becoming an official IndigoBox (/Pentelec) project that I
work on at three in the morning.

Thank you for visiting, and may this be the start of an ever-developing,
fun, and successful journey!

(This text is also going to be copied to readme.md...shortly.)