Added client side files and resources
What are these, you ask? This is what gets installed on a user's
machine...the actual files. That's right; if you want to actually run a
version of Project Snowshoes without compiling the code that is in the
"ProjectSnowshoes" directory, you can use this. Actually, you need these
either way just to run the program, as the file locations are directly
called in the program. Hopefully, that will change soon to be a bit more
flexible...but we shall see here.

IMPORTANT: Read the readme.md file at section "What am I looking at?"
for installation instructions.