Removal of Application.DoEvents() from NewAccount-New2
While it does allow for the form to be repainted in the middle of
copying files, the usage of DoEvents() and its ability to let more than
just the next OnPaint() be carried through causes unacceptable wait
time.

This also has a lot of changes from me trying to run this on an "Any
CPU" setting, having to update a lot of files only to not compile due to
build errors here. Hopefully I can make this not x64-specific, but if
that's our only concern, I'm okay with it for now.