Required Accent Color Selection Repair
This is more of a bug fix, as the program would eventually crash without
having a color selected. This commit adds extra prompts for the user to
make sure they select a color, restricting them from moving on before
they do so.

I had a much more elegant solution for this originally, but something is
wrong with my logic at 0610 hours, and I should probably just go to
sleep now.