Bug Fix: Properly opening and closing ChangeBackground color panel
A previous bug fix sought to make sure one could not spawn multiple
instances of the ChangeBackground color panel, but if a user wanted to
change the color they picked, they could no longer open up the panel
properly. Luckily, I set this up for myself very nicely (meaning I was
actually sane for once), and everything was dependent on a Boolean
value.