Bug Fixes: Removed ability to have multiple Minerva windows, and disabled Speech Recognition
This seems like I am disabling features, yet the latter relies on a COM
object that is not only throwing an exception when attempting
generation, yet does not work to the extent we wish it to when actually
started, anyway. The former, on the other hand, is to prevent users from
clicking on the Minerva button and having multiple instances of Minerva
that are actually just overlaying on top of each other (therefore, you
cannot see two or more at the same time...unless Explorer is still
active and you can read the render of the open forms while reading the
current Minerva form that is still open).