Switch to .NET Framework 4.6, and Creation of app.manifest
In an effort to make text reproduction cleaner, I did some research and
actually was able to produce some nice-looking results. Unfortunately,
scaling issues that I subverted before through writing my own code now
is coming back to bite, and at this point, it's better off if we keep
them and re-align our focuses in the next build of PS. That said, this
all ushered in the adoption of .NET Framework 4.6, and there is finally
an application manifest to toy around with!

If things break, this will be the commit that broke them.