Removed hidden files from listing on Space
This can be an option to add in later, but the space is meant to
highlight important items...System files like desktop.ini or the
AlbumArtSmall image might not be part of that list. They still mean a
lot to us, so they are not touched and remain in the directory if they
exist here.