Event Handling for UserAcc Text
Look at this. This is pathetic. Alas, when you get only 90 seconds in
the time you originally devoted (but had taken up...that was 5 hours!!)
to set up and write anything, this is the result. I haven't even been
able to compile it yet.

Here's hoping to sneaking in another commit before the night ends here.