Removed AccountInfo.cs, Checked for Custom Fonts
The AccountInfo.cs form was the one I created originally for the...well,
account information page. Turns out, I just kept it on UserSettings, so
no need for that form anymore. The custom fonts are the only thing that
I could see being customized outside of Snowshoes (by hard-coding a
property, or doing something else), so I threw those in when checking
for custom properties.