Attempt to Resolve Window Handling: Part One
This is a very straightforward commit, and hey, even has some code that
I had to thank somebody for and basically cleans things out. This is
something that I have not worked with in a very, very long time, so this
will take a while to publish this fix here. I apologize.