Handling Handles to Report Multiple Windows
"Lol this in a way works tho" - Me on first run.

Now every title is producing at least one Default IME reporting...but
we'll fix that. No worries.

Just kidding, everybody run. This is becoming more and more of a cluster
as this develops. That said, multiple windows are reported now!...it's
just that they overlay on top of each other. We've got some work to do
here.