How is this working?
Doges of the world, I learned a lot about window management again today.
This was really fun...and surprisingly more concise than expected. Note
that the correct icons are still not being returned, but this is still
something that makes me very happy...I've been waiting to do this for a
very long time.