Unfinished: Win + D to return to Space
This is the first commit off a T420, so it's been fun. I'll be using it
for more, shall we say, experimental things.

Anyway, so Windows Key + D is one of the many key combinations that
breaks without Explorer open. This makes sense in versions of Windows
like 8 and higher, because the entire desktop down to the wallpaper is
even controlled by Explorer (as well as some Metro styles...try Alt+Tab
in Snowshoes).

That reminds me, I should probably test in 7 soon.

ONE MORE THING: With the addition of UAC needed to grant administrator
access, this (as far as I know) no longer supports Windows XP.