Background Workers...Work!
Progress when copying files is finally reported, so that you aren't left
out in the dark during that process wondering if Snowshoes froze. I
mean, that's still a possibility, but a lot less likely now.

Also fixed Space default location to pull from Desktop, and also
introduced a check for # forms open so that exiting setup would know
whether to close the application entirely or not.